using uints

+++ src/Stratis.Bitcoin.Features.Consensus/Rules/CommonRules/CheckDifficultyPosRule.cs
@@ -21,13 +21,19 @@ namespace Stratis.Bitcoin.Features.Consensus.Rules.CommonRules
                 return;
             }
 
+            // TODO: In the future once we migrated to fully C# network it might be good to consider signaling in the block header the network type.
+
+            // In order to calculate difficulty we need to know the if the block type is POW/POS.
+            // This is only available when the block is downloaded (on the coinbase).
+            // to validate headers before having the block itself the best we can do is
+            // validate POS header after POW blocks era is finished.
             if (context.ValidationContext.ChainedHeader.Height + 2 > this.Parent.Network.Consensus.LastPOWBlock)
             {
                 ChainedHeader chainedHeader = context.ValidationContext.ChainedHeader;
                 BlockHeader first = chainedHeader.Previous.Header;
                 BlockHeader second = chainedHeader.Previous.Previous.Header;
 
-                Target nextWorkRequired = this.PosParent.StakeValidator.CalculateRetarget((int)first.Time, first.Bits, (int)second.Time, context.Consensus.ProofOfStakeLimitV2);
+                Target nextWorkRequired = this.PosParent.StakeValidator.CalculateRetarget(first.Time, first.Bits, second.Time, context.Consensus.ProofOfStakeLimitV2);
 
                 BlockHeader header = context.ValidationContext.Block.Header;
